refactor: split storage/mod.rs into domain sub-files
storage/mod.rs (844 lines) split into focused files:
- config.rs   — admin and fee helpers (instance storage)
- vault.rs    — vault metadata helpers
- issuer.rs   — authorized and denied issuer index (O(1) swap-and-pop)
- credential.rs — VC payloads, O(1) VC index, parent links, VC status
- ttl.rs      — TTL extension helpers
- sponsor.rs  — sponsored vault config

mod.rs becomes a ~60-line hub: DataKey enum + pub use re-exports.
All callers continue to use storage::* unchanged.
